# backend/openapi.yaml
openapi: 3.1.0
info:
  title: itgel API
  version: 1.0.0
  description: |
    Захиалгын дэлгүүрийн backend API.

    - Бүх мөнгөн дүн бүхэл тоо (₮).
    - Бүх огноо ISO 8601, `Asia/Ulaanbaatar` (UTC+8) цагаар бодогдоно.
    - Хэрэглэгчийн API `costPrice`-г хэзээ ч буцаахгүй.
servers:
  - url: http://localhost:4000
    description: Dev
...
security: []
paths:
...
  /api/admin/customers:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Хэрэглэгчид — захиалгын тоо, нийт зарцуулалт, сүүлд
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        '200': { description: Жагсаалт }
  /api/admin/customers/{id}:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Хэрэглэгчийн дэлгэрэнгүй, захиалгын түүх
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Хэрэглэгч }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Манай JWT (`/api/auth/verify` эсвэл `/api/admin/auth/login`) эсвэл
        Supabase Auth-ийн access token (JWKS-ээр шалгагдана).
...